Move private helper structs out of WFClient file
